+++ core/common/dmstream.cpp
@@ -7,7 +7,7 @@ DynamicMemoryStream::DynamicMemoryStream()
 {
 }
 
-void  DynamicMemoryStream::checkSize(int size)
+void  DynamicMemoryStream::checkSize(size_t size)
 {
     if (size > m_buffer.size())
         while (m_buffer.size() < size)
@@ -20,7 +20,7 @@ int  DynamicMemoryStream::read(void *buf, int count)
         throw StreamException("End of stream");
 
     auto end = std::copy(m_buffer.begin() + m_pos,
-                         m_buffer.begin() + std::min(m_pos + count, (int) m_buffer.size()),
+                         m_buffer.begin() + std::min(m_pos + count, m_buffer.size()),
                          static_cast<char*>(buf));
     auto bytesRead = end - static_cast<char*>(buf);
     m_pos += bytesRead;
