+++ core/common/http.cpp
@@ -103,7 +103,7 @@ bool    HTTP::nextHeader()
                 if (!(end = strchr(ap, '\n')))
                     end = ap + strlen(ap);
             value = string(ap, end);
-            for (int i = 0; i < name.size(); ++i)
+            for (size_t i = 0; i < name.size(); ++i)
                 name[i] = toupper(name[i]);
             headers.set(name, value);
         }
